address comment

+++ components/raftstore/src/coprocessor/region_info_accessor.rs
@@ -690,7 +690,7 @@ impl RegionCollector {
             s.cop_detail.iterated_count() >= iterated_count_to_filter
                 // plus processed_keys by 1 to make it not 0
                 && s.cop_detail.mvcc_amplification()
-                    >= self.mvcc_amplification_threshold
+                    >= self.mvcc_amplification_threshold as f64
         });
 
         // TODO(SpadeA): remove it when auto load/evict is stable
